+++ packages/ricos-content/web/src/velo-adapter/select.spec.ts
@@ -1,14 +1,17 @@
 import migrationContent from '../../statics/json/migratedFixtures/migration-content.json';
-import { query } from './query';
-import type { RichContent, Node } from 'ricos-schema';
+import rawContent from '../../tests/modifyFixtures/images-dividers.json';
+import { select } from './select';
+import type { Node } from 'ricos-schema';
 import {
   Node_Type,
   PluginContainerData_Alignment,
   PluginContainerData_Width_Type,
   Decoration_Type,
   FontSizeData_fontType,
+  RichContent,
 } from 'ricos-schema';
 import type { TextNode, VideoNode } from '../types/node-refined-types';
+import { extract } from '../RicosContentAPI/extract';
 
 const videoUrl = 'https://www.youtube.com/watch?v=CoJ23XNHgG0';
 
@@ -78,27 +81,27 @@ describe('query', () => {
   const content = migrationContent as unknown as RichContent;
 
   it('should return all results if no filtering was done', () => {
-    const result = query({ nodes: [videoMock] }).find();
+    const result = select({ nodes: [videoMock] }).find();
     expect(result[0].nodes).toStrictEqual<Node[]>([videoMock]);
   });
 
   it('should do eq', () => {
-    const result = query(content).eq('videoData.video.src.url', videoUrl).find();
+    const result = select(content).eq('videoData.video.src.url', videoUrl).find();
     expect(result).toStrictEqual<VideoNode[]>([videoMock]);
   });
 
   it('should do gt', () => {
-    const result = query(content).gt('videoData.thumbnail.width', 490).find();
+    const result = select(content).gt('videoData.thumbnail.width', 490).find();
     expect(result).toStrictEqual<VideoNode[]>([videoMock_BiggerWidth]);
   });
 
   it('should do lt', () => {
-    const result = query(content).lt('videoData.thumbnail.width', 490).find();
+    const result = select(content).lt('videoData.thumbnail.width', 490).find();
     expect(result).toStrictEqual<VideoNode[]>([videoMock]);
   });
 
   it('should do isEmpty - should keep relevant fields', () => {
-    const result = query(content)
+    const result = select(content)
       .eq('videoData.video.src.url', videoUrl)
       .isEmpty('some.dummy.field.that.should.not.exist')
       .find();
@@ -106,7 +109,7 @@ describe('query', () => {
   });
 
   it('should do isEmpty - should filter for existing fields', () => {
-    const result = query(content)
+    const result = select(content)
       .eq('videoData.video.src.url', videoUrl)
       .isEmpty('videoData.video.src.url')
       .find();
@@ -114,7 +117,7 @@ describe('query', () => {
   });
 
   it('should do isNotEmpty - should remove inexistent fields', () => {
-    const result = query(content)
+    const result = select(content)
       .eq('videoData.video.src.url', videoUrl)
       .isNotEmpty('some.dummy.field.that.should.not.exist')
       .find();
@@ -122,7 +125,7 @@ describe('query', () => {
   });
 
   it('should do isNotEmpty - should keep existing fields', () => {
-    const result = query(content)
+    const result = select(content)
       .eq('videoData.video.src.url', videoUrl)
       .isNotEmpty('videoData.video.src.url')
       .find();
@@ -130,27 +133,58 @@ describe('query', () => {
   });
 
   it('should do startsWith', () => {
-    const result = query(content).startsWith('textData.text', 'Lorem ipsum dolor').find();
+    const result = select(content).startsWith('textData.text', 'Lorem ipsum dolor').find();
     expect(result).toStrictEqual<TextNode[]>([textNodeMock]);
   });
 
   it('should do endsWith', () => {
-    const result = query(content).endsWith('textData.text', 'sit amet').find();
+    const result = select(content).endsWith('textData.text', 'sit amet').find();
     expect(result).toStrictEqual<TextNode[]>([textNodeMock]);
   });
 
   it('should do limit', () => {
-    const result = query(content).limit(2).find();
+    const result = select(content).limit(2).find();
     expect(result.length).toStrictEqual<number>(2);
   });
 
   it('should do count', () => {
-    const result = query(content).count();
+    const result = select(content).count();
     expect(result).toStrictEqual<number>(186);
   });
 
   it('should do count + limit', () => {
-    const result = query(content).limit(2).count();
+    const result = select(content).limit(2).count();
     expect(result).toStrictEqual<number>(2);
+  });
+});
+
+describe('modify', () => {
+  it('should trigger callback', () => {
+    const content = RichContent.fromJSON(rawContent);
+    const callback = jest.fn();
+    const _ = select(content, callback)
+      .eq('type', Node_Type.DIVIDER)
+      .modify(n => [n, n]);
+
+    expect(callback).toBeCalledTimes(1);
+  });
+  it('velo modifier should work (basic test)', () => {
+    const content = RichContent.fromJSON(rawContent);
+    const callback = jest.fn();
+
+    // multiply dividers
+    const result = select(content, callback)
+      .eq('type', Node_Type.DIVIDER)
+      .modify(n => [n, n]);
+
+    const extractDividers = (content: RichContent) =>
+      extract(content.nodes)
+        .filter(node => node.type === Node_Type.DIVIDER)
+        .get();
+
+    // check total amount of dividers has been doubled
+    const originalDividersCount = extractDividers(content).length;
+    const newDividersCount = extractDividers(result).length;
+    expect(originalDividersCount * 2).toBe(newDividersCount);
   });
 });